Show sticky notification on app start
